+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 ## [À venir dans la prochaine version]
 ### Ajouté
 - Permettre à de nouveaux administrateurs de s'inscrire et d'ajouter leur société
+- Ajout de l'activité "X à modifié le fait marquant Y sur le projet Z"